updateIcon: Respect system colour scheme
Currently the extension toolbar icon does not respect the selected colour scheme (light/dark mode). This causes the light-mode icon to be used on a dark mode toolbar, thus making the icon almost unreadable.

This change implements a simple check that picks the correct icon based on the system's colour scheme instead.